Refactor screen styles to use shared top level stylesheet reducing code
