+++ src/screens/styles.js
@@ -1,5 +1,5 @@
 import { StyleSheet } from "react-native";
-import { white } from "../../constants";
+import { white } from "../constants";
 
 export default StyleSheet.create({
   container: {
@@ -13,5 +13,8 @@ export default StyleSheet.create({
     width: "100%",
     paddingHorizontal: 20,
     paddingVertical: 10
+  },
+  input: {
+    marginVertical: 10
   }
 });
